refact: alteração nas tabelas do banco e na autenticação

+++ api-js/src/modules/auth/auth.service.js
@@ -1,3 +1,4 @@
+const UAParser = require("ua-parser-js");
 const repository = require("./auth.repository");
 const authUtil = require("../../common/utils/auth.util");
 const tokenUtil = require("../../common/utils/token.util");
@@ -10,22 +11,14 @@ const signUp = async ({ name, email, password, userAgent, ipAddress }) => {
 
   hashedPassword = await authUtil.hashPassword(password);
   const user = await repository.createUser({ name, email, password: hashedPassword });
+  const userId = user.id;
 
-  ipAddress = authUtil.normalizeIp(ipAddress);
+  let { accessToken, refreshToken } = tokenUtil.createTokens({ userId, userAgent });
 
-  const { accessToken, refreshToken } = tokenUtil.createTokens({
-    userId: user.id,
-    userAgent,
-    ipAddress,
-  });
+  const expiresAt = new Date(tokenUtil.decodeJWT(refreshToken).exp * 1000);
+  refreshToken = cryptoUtil.encryptData(refreshToken);
 
-  return await repository.createSession({
-    userId: user.id,
-    accessToken,
-    refreshToken,
-    userAgent,
-    ipAddress,
-  });
+  return await repository.createSession({ userId, accessToken, refreshToken, userAgent, ipAddress, expiresAt });
 };
 
 const signIn = async ({ email, password, userAgent, ipAddress }) => {
@@ -35,70 +28,57 @@ const signIn = async ({ email, password, userAgent, ipAddress }) => {
     throw new CustomError("Credenciais inválidas", 401);
   }
 
-  ipAddress = authUtil.normalizeIp(ipAddress);
+  const userId = user.id;
+  const ua = UAParser(userAgent);
 
-  console.log("👷[SERVICE] - userAgent: ", userAgent);
-  console.log("👷[SERVICE] - ipAddress: ", ipAddress);
+  const sessionData = {
+    ip: ipAddress,
+    device: `${ua.device.vendor || ""} ${ua.device.model || ""}`.trim(),
+    os: `${ua.os.name || ""} ${ua.os.version || ""}`.trim(),
+    browser: `${ua.browser.name || ""} ${ua.browser.version || ""}`.trim(),
+  };
 
-  const existingSession = await repository.findSession({ userId: user.id, userAgent, ipAddress });
-  if (existingSession) await repository.deleteSession(existingSession.id);
+  const existingSession = await repository.getSessionByUserId({ userId, userAgent });
+  if (existingSession) await repository.deleteSessionByAgent({ userId, userAgent });
 
-  const { accessToken, refreshToken } = tokenUtil.createTokens({
-    userId: user.id,
-    userAgent,
-    ipAddress,
-  });
+  let { accessToken, refreshToken } = tokenUtil.createTokens({ userId, userAgent });
 
-  return await repository.createSession({
-    userId: user.id,
-    accessToken,
-    refreshToken,
-    userAgent,
-    ipAddress,
-  });
+  const expiresAt = new Date(tokenUtil.decodeJWT(refreshToken).exp * 1000);
+  refreshToken = cryptoUtil.encryptData(refreshToken);
+
+  return await repository.createSession({ userId, accessToken, refreshToken, userAgent, ipAddress, expiresAt });
 };
 
-const signOut = async ({ userId, refreshToken, userAgent, ipAddress }) => {
-  if (!refreshToken) throw new CustomError("Token de sign out não fornecido", 404);
-
-  const session = await repository.findSession({ userId, refreshToken, userAgent, ipAddress });
-
-  if (session) {
-    await repository.deleteSession(session.id);
-  }
+const signOut = async ({ userId, userAgent }) => {
+  const session = await repository.getSessionByUserId({ userId, userAgent });
+  if (session) await repository.deleteSession(session.id);
   return true;
 };
 
-const revalidateTokens = async ({ refreshToken, userAgent, ipAddress }) => {
-  if (!refreshToken) {
-    throw new CustomError("Token de atualização nao fornecido", 401);
-  }
+const revalidateTokens = async ({ req, refreshToken, userAgent, ipAddress }) => {
+  if (!refreshToken) throw new CustomError("Token de atualização nao fornecido", 401);
 
   const decrypted = cryptoUtil.decryptData(refreshToken);
   const { userId } = tokenUtil.decodeJWT(decrypted);
 
-  const session = await repository.findSession({ userId, refreshToken, userAgent, ipAddress });
-
-  console.log("👷 [SERVICE] - session: ", `UID: ${userId}`, `RFT: ${refreshToken}`, `UAG: ${userAgent}`, `IP: ${ipAddress}`);
+  const session = await repository.getSessionByUserId({ userId, userAgent });
 
   if (!session || session.expiresAt < new Date()) {
-    if (session) await deleteSession(session.id);
+    if (session) await repository.deleteSession(req, session.id);
     throw new CustomError("Sessão inválida ou expirada", 401);
   }
 
-  const newTokens = tokenUtil.createTokens({ userId, userAgent, ipAddress });
+  const newTokens = tokenUtil.createTokens({ userId, userAgent });
+  const expiresAt = new Date(tokenUtil.decodeJWT(newTokens.refreshToken).exp * 1000);
 
   return await repository.createSession({
     userId,
     accessToken: newTokens.accessToken,
-    refreshToken: newTokens.refreshToken,
+    refreshToken: cryptoUtil.encryptData(newTokens.refreshToken),
     userAgent,
     ipAddress,
+    expiresAt,
   });
 };
 
-const getSessions = async (userId) => {
-  return await repository.getSessions(userId);
-};
-
-module.exports = { signUp, signIn, signOut, revalidateTokens, getSessions };
+module.exports = { signUp, signIn, signOut, revalidateTokens };
